Working web app with encode and decode without encryption
